# [improve][deploy] changed the default GC options to ZGC from G1GC